LPS-99142 Moving logic to the base class to use it in folders too
