+++ modules/apps/document-library/document-library-web/src/main/java/com/liferay/document/library/web/internal/portlet/action/GetFolderMVCRenderCommand.java
@@ -20,11 +20,20 @@ import com.liferay.document.library.web.internal.util.DLTrashUtil;
 import com.liferay.portal.kernel.portlet.bridges.mvc.MVCRenderCommand;
 import com.liferay.portal.kernel.security.auth.PrincipalException;
 import com.liferay.portal.kernel.servlet.SessionErrors;
+import com.liferay.portal.kernel.util.GetterUtil;
+import com.liferay.portal.kernel.util.Portal;
+import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.kernel.util.WebKeys;
 
 import javax.portlet.PortletException;
+import javax.portlet.PortletRequest;
 import javax.portlet.RenderRequest;
 import javax.portlet.RenderResponse;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpSession;
+
+import org.osgi.service.component.annotations.Reference;
 
 /**
  * @author Iván Zaera
@@ -43,6 +52,16 @@ public abstract class GetFolderMVCRenderCommand implements MVCRenderCommand {
 
 			renderRequest.setAttribute(
 				DLWebKeys.DOCUMENT_LIBRARY_TRASH_UTIL, getDLTrashUtil());
+
+			HttpSession session = _getPortalSession(renderRequest);
+
+			String error = GetterUtil.getString(session.getAttribute("error"));
+
+			if (Validator.isNotNull(error)) {
+				session.removeAttribute("error");
+
+				SessionErrors.add(renderRequest, error);
+			}
 		}
 		catch (Exception e) {
 			if (e instanceof NoSuchFolderException ||
@@ -62,5 +81,16 @@ public abstract class GetFolderMVCRenderCommand implements MVCRenderCommand {
 	protected abstract DLTrashUtil getDLTrashUtil();
 
 	protected abstract String getPath();
+
+	@Reference
+	protected Portal portal;
+
+	private HttpSession _getPortalSession(PortletRequest portletRequest) {
+		HttpServletRequest originalHttpServletRequest =
+			portal.getOriginalServletRequest(
+				portal.getHttpServletRequest(portletRequest));
+
+		return originalHttpServletRequest.getSession();
+	}
 
 }
